+++ modules/AWSPowerShell/Cmdlets/EC2/Basic/Get-EC2ClassicLinkInstance-Cmdlet.cs
@@ -33,8 +33,8 @@ namespace Amazon.PowerShell.Cmdlets.EC2
     /// use this request to return information about other instances.
     /// 
     ///  <note><para>
-    /// We are retiring EC2-Classic on August 15, 2022. We recommend that you migrate from
-    /// EC2-Classic to a VPC. For more information, see <a href="https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/vpc-migrate.html">Migrate
+    /// We are retiring EC2-Classic. We recommend that you migrate from EC2-Classic to a VPC.
+    /// For more information, see <a href="https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/vpc-migrate.html">Migrate
     /// from EC2-Classic to a VPC</a> in the <i>Amazon Elastic Compute Cloud User Guide</i>.
     /// </para></note><br/><br/>This cmdlet automatically pages all available results to the pipeline - parameters related to iteration are only needed if you want to manually control the paginated output. To disable autopagination, use -NoAutoIteration.
     /// </summary>