+++ modules/AWSPowerShell/Cmdlets/EC2/Basic/Get-EC2FlowLog-Cmdlet.cs
@@ -28,9 +28,14 @@ using Amazon.EC2.Model;
 namespace Amazon.PowerShell.Cmdlets.EC2
 {
     /// <summary>
-    /// Describes one or more flow logs. To view the information in your flow logs (the log
-    /// streams for the network interfaces), you must use the CloudWatch Logs console or the
-    /// CloudWatch Logs API.<br/><br/>This cmdlet automatically pages all available results to the pipeline - parameters related to iteration are only needed if you want to manually control the paginated output. To disable autopagination, use -NoAutoIteration.
+    /// Describes one or more flow logs.
+    /// 
+    ///  
+    /// <para>
+    /// To view the published flow log records, you must view the log destination. For example,
+    /// the CloudWatch Logs log group, the Amazon S3 bucket, or the Kinesis Data Firehose
+    /// delivery stream.
+    /// </para><br/><br/>This cmdlet automatically pages all available results to the pipeline - parameters related to iteration are only needed if you want to manually control the paginated output. To disable autopagination, use -NoAutoIteration.
     /// </summary>
     [Cmdlet("Get", "EC2FlowLog")]
     [OutputType("Amazon.EC2.Model.FlowLog")]
@@ -46,9 +51,8 @@ namespace Amazon.PowerShell.Cmdlets.EC2
         /// <summary>
         /// <para>
         /// <para>One or more filters.</para><ul><li><para><code>deliver-log-status</code> - The status of the logs delivery (<code>SUCCESS</code>
-        /// | <code>FAILED</code>).</para></li><li><para><code>log-destination-type</code> - The type of destination to which the flow log
-        /// publishes data. Possible destination types include <code>cloud-watch-logs</code> and
-        /// <code>s3</code>.</para></li><li><para><code>flow-log-id</code> - The ID of the flow log.</para></li><li><para><code>log-group-name</code> - The name of the log group.</para></li><li><para><code>resource-id</code> - The ID of the VPC, subnet, or network interface.</para></li><li><para><code>traffic-type</code> - The type of traffic (<code>ACCEPT</code> | <code>REJECT</code>
+        /// | <code>FAILED</code>).</para></li><li><para><code>log-destination-type</code> - The type of destination for the flow log data
+        /// (<code>cloud-watch-logs</code> | <code>s3</code> | <code>kinesis-data-firehose</code>).</para></li><li><para><code>flow-log-id</code> - The ID of the flow log.</para></li><li><para><code>log-group-name</code> - The name of the log group.</para></li><li><para><code>resource-id</code> - The ID of the VPC, subnet, or network interface.</para></li><li><para><code>traffic-type</code> - The type of traffic (<code>ACCEPT</code> | <code>REJECT</code>
         /// | <code>ALL</code>).</para></li><li><para><code>tag</code>:&lt;key&gt; - The key/value combination of a tag assigned to the
         /// resource. Use the tag key in the filter name and the tag value as the filter value.
         /// For example, to find all resources that have a tag with the key <code>Owner</code>