+++ modules/AWSPowerShell/Cmdlets/SageMaker/Basic/New-SMTrainingJob-Cmdlet.cs
@@ -45,7 +45,12 @@ namespace Amazon.PowerShell.Cmdlets.SM
     /// optimize this learning process. For a list of hyperparameters for each training algorithm
     /// provided by SageMaker, see <a href="https://docs.aws.amazon.com/sagemaker/latest/dg/algos.html">Algorithms</a>.
     /// 
-    /// </para></li><li><para><code>InputDataConfig</code> - Describes the training dataset and the Amazon S3,
+    /// </para><important><para>
+    /// You must not include any security-sensitive information, such as account access IDs,
+    /// secrets, and tokens, in the dictionary for configuring hyperparameters. SageMaker
+    /// rejects the training job request and returns an exception error for detected credentials,
+    /// if such user input is found.
+    /// </para></important></li><li><para><code>InputDataConfig</code> - Describes the training dataset and the Amazon S3,
     /// EFS, or FSx location where it is stored.
     /// </para></li><li><para><code>OutputDataConfig</code> - Identifies the Amazon S3 bucket where you want SageMaker
     /// to save the results of model training. 
@@ -209,7 +214,10 @@ namespace Amazon.PowerShell.Cmdlets.SM
         /// algorithm provided by SageMaker, see <a href="https://docs.aws.amazon.com/sagemaker/latest/dg/algos.html">Algorithms</a>.
         /// </para><para>You can specify a maximum of 100 hyperparameters. Each hyperparameter is a key-value
         /// pair. Each key and value is limited to 256 characters, as specified by the <code>Length
-        /// Constraint</code>. </para>
+        /// Constraint</code>. </para><important><para>You must not include any security-sensitive information, such as account access IDs,
+        /// secrets, and tokens, in the dictionary for configuring hyperparameters. SageMaker
+        /// rejects the training job request and returns an exception error for detected credentials,
+        /// if such user input is found.</para></important>
         /// </para>
         /// </summary>
         [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]